% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_classification.R
\name{explain_classification}
\alias{explain_classification}
\title{Explain how each row was classified according to the rules}
\usage{
explain_classification(data, ruleset, outcome_col_name = "OUTCOME")
}
\arguments{
\item{data}{A data frame that was already classified by \code{\link{apply_rules}}, i.e.,
it has a final outcome column.}

\item{ruleset}{An object of class "academic_ruleset".}

\item{outcome_col_name}{The name of the classification column in \code{data} (e.g. "OUTCOME").}
}
\value{
A list of length \code{nrow(data)}, where each element has:
\itemize{
\item \code{index}: row index
\item \code{final_classification}: the final classification (PASS or FAIL, etc.)
\item \code{rules_evaluation}: a list of texts describing whether each rule was triggered
}
}
\description{
For each row, indicates whether each rule triggered a fail or not, and shows the final outcome.
}
\examples{
df <- data.frame(id=1:2, score=c(60,45), fails=c(0,3))
rs <- define_rule_set(list(
  min_score = list(type="minimum", value=50, dimension="score"),
  max_fails = list(type="maximum", value=2, dimension="fails")
))

out <- apply_rules(df, rs, "OUTCOME")
explain_classification(out, rs, "OUTCOME")
}
